feat: rename and restructure the map lib classes, #36
