--- src/lib/maps/basemap/index.js
@@ -1,5 +1,0 @@
-import BaseMap from './basemap'
-
-export {
-  BaseMap
-}
