--- src/lib/maps/drawing/index.js
@@ -1,5 +1,0 @@
-import MapDraw from './mapdraw'
-
-export {
-  MapDraw
-}
