+++ src/lib/maps/google/index.js
@@ -1,9 +1,9 @@
 import GoogleMap from './basic'
-import GoogleMapLeaflet from './mutant'
+import LeafletGoogleMap from './mutant'
 import GoogleMapDraw from './drawing'
 
 export {
   GoogleMap,
-  GoogleMapLeaflet,
+  LeafletGoogleMap,
   GoogleMapDraw
 }
