+++ src/lib/maps/leaflet/basic/basic.js
@@ -1,15 +1,15 @@
-import { MapUtils } from '../common'
+import { MapUtils } from '../../common'
 
 /**
  * Base class for managing LeafletJS web maps.
  */
-class BaseMap extends MapUtils {
+class LeafletMap extends MapUtils {
   map
   mapId
   config
 
   /**
-   * BaseMap constructor parameters
+   * LeafletMap constructor parameters
    * @typedef {Object} config
    * @param {String} config.mapId - HTML DOM id where to render the map.
    * @param {String} config.lat - Latitude.
@@ -51,4 +51,4 @@ class BaseMap extends MapUtils {
   }
 }
 
-export default BaseMap
+export default LeafletMap
