+++ src/lib/maps/leaflet/basic/index.js
@@ -1,0 +1,5 @@
+import LeafletMap from './basic'
+
+export {
+  LeafletMap
+}
