+++ src/lib/maps/leaflet/drawing/index.js
@@ -1,0 +1,5 @@
+import LeafletMapDraw from './leafletmapdraw'
+
+export {
+  LeafletMapDraw
+}
