--- src/lib/maps/mapbox/index.js
@@ -1,5 +1,0 @@
-import WebMapBox from './mapboxmap'
-
-export {
-  WebMapBox
-}
